refactor: split CLI theme command
